# docs/swagger-ui/openapi-gateway.yaml
openapi: 3.0.0
info:
  title: Sports Bet Platform - API Gateway
  version: 1.0.0
servers:
  - url: http://localhost:8000
...
paths:
...
  /api/wallet/wallet:
    get:
      tags: [Wallet]
      summary: Consulta saldo da carteira do usuário
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dados da carteira
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
...
  /api/bets/bets:
    post:
      tags: [Bets]
      summary: Cria uma nova aposta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceBetRequest'
      responses:
        '200':
          description: Aposta criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceBetResponse'
...
components:
  schemas:
...
    WalletResponse:
      type: object
      properties:
        userId: { type: string }
        walletId: { type: string }
        balance_cents: { type: integer }
...
    PlaceBetRequest:
      type: object
      properties:
        userId: { type: string }
        eventId: { type: string }
        market: { type: string }
        selection: { type: string }
        stake_cents: { type: integer }
        odd_value: { type: number }
      required: [userId, eventId, market, selection, stake_cents, odd_value]
    PlaceBetResponse:
      type: object
      properties:
        betId: { type: string }
        status: { type: string }
        new_balance: { type: integer }
        message: { type: string }
...